requirements.txt
websockets>=12.0
aiofiles>=23.0
pyyaml>=6.0
